+++ khairullin.vadim/P3/functions.hpp
@@ -7,8 +7,8 @@ namespace khairullin
   int check_arguments(int argc, char ** argv, int command);
   void filling(int * array, std::ifstream & input, size_t rows, size_t cols, int command);
   int localMax(const int * array, size_t rows, size_t cols);
-  int CheckZero(const int * array, int index1, int index2, int size);
-  bool triangle(const int * array, int size);
+  int CheckZero(const int * array, size_t index1, size_t index2, size_t size);
+  bool triangle(const int * array, size_t size);
 }
 
 #endif
